+++ eras/babbage/impl/CHANGELOG.md
@@ -13,6 +13,10 @@
 * Expose `fixupCollateralReturn`
 * Add `DecCBOR` instance for `Block`
 
+### `cddl`
+
+* Add `HuddleRule "vrf_cert"` instance
+
 ## 1.14.0.0
 
 * Add `Ord` instances for `BabbageContextError`, `BabbageTxOut`, `BabbageUtxoPredFailure`, `BabbageUtxowPredFailure`